Create config Azure Function
