+++ azure-function/yield-curve/__init__.py
@@ -1,7 +1,8 @@
 import logging
 import pandas as pd
-
 import azure.functions as func
+
+from ..shared_code import parameter
 
 input_data = pd.DataFrame({
     'Currency': ['USD'] * 2 + ['CAD'] * 2,
@@ -9,6 +10,8 @@ input_data = pd.DataFrame({
     'Rate': [0.0250, 0.0350, 0.0245, 0.0256],
     'Maturity': [2.0, 3.0, 2.0, 3.0]
 })
+
+
 def main(req: func.HttpRequest, rates) -> func.HttpResponse:
     """Function for API call
 
@@ -23,25 +26,13 @@ def main(req: func.HttpRequest, rates) -> func.HttpResponse:
     """
     logging.info('Python HTTP trigger function processed the rates API request.')
 
-    param = get_param(req=req, param="data_format")
+    param = parameter.get_param(req=req, param="data_format")
 
     if param == "csv":
         rates_df = pd.read_json(rates)
-        return func.HttpResponse(rates_df.to_csv(), mimetype="text/csv", headers={"Content-disposition":
-       "attachment; filename=eiopa_rfr_yield_curves.csv"})
+        return func.HttpResponse(rates_df.to_csv(),
+                                 mimetype="text/csv",
+                                 headers={"Content-disposition": "attachment; filename=eiopa_rfr_yield_curves.csv"})
     else:
         return func.HttpResponse(rates, mimetype="application/json")
 
-
-def get_param(req: func.HttpRequest, param: str):
-
-    param_value = req.params.get(param)
-    if not param_value:
-        try:
-            req_body = req.get_json()
-        except ValueError:
-            pass
-        else:
-            param_value = req_body.get(param)
-
-    return param_value
